+++ lib/torrent_tracker.dart
@@ -1,3 +1,4 @@
+/// Bittorrent Tracker. Include UDP tracker and Http/Https tracker
 library torrent_tracker;
 
 export 'src/tracker/tracker_base.dart';